Remove duplicate literal values #9  Fix
IfcConvertor.resource_map was created to take care of duplicate resource
entries. As the string list is not used anywhere, it was replaced by a
list to save some memory. IfcConvertor.fillProperties was restructured
to to manage the modifications. The result was validated and it seems to
be OK. Validation was passed OK.  "generated RDF graph is OK! Writing
TTL and RDF file..."